fix: use tracing for event logs, fix u8 overflow
Replace println!/print! with tracing::info! so all output goes
to stderr, avoiding [NB blob data] in systemd-journal. Fix
max_contacts calculation overflow when payload[2] > 127.

+++ src/protocol/decode.rs
@@ -332,7 +332,7 @@ pub fn decode_response_payload(code: u8, payload: &[u8]) -> Option<HashMap<Strin
             if payload.len() > 60 {
                 map.insert(
                     "max_contacts".into(),
-                    DecodedValue::Integer((payload[2] * 2) as i64),
+                    DecodedValue::Integer((payload[2] as u16 * 2) as i64),
                 );
                 map.insert(
                     "max_channels".into(),
